refactored main functions, made all solution functions return String

+++ lib/src/lib.rs
@@ -14,7 +14,7 @@ use std::fs;
 ///
 /// # Panics
 /// If no input file is present
-pub fn wrapper<F: FnOnce(&str) -> T, T: Display>(year: u16, index: u8, part: u8, func: F) {
+pub fn wrapper<F: FnOnce(&str) -> T, T: Display>(year: u16, index: usize, part: u8, func: F) {
     println!(
         "Day {index} part {part}: {}",
         func(&fs::read_to_string(format!("./ac{year}/input/p{index:0>2}")).unwrap())
